moved snippet to plugin folder
