+++ lib/ejs.js
@@ -467,10 +467,6 @@ exports.clearCache = function () {
   exports.cache.reset();
 };
 
-exports.enableSnippets = function () {
-  require('./ejs-snippet');
-};
-
 function Template(text, opts) {
   opts = opts || {};
   var options = {};
